Fix issue #70
Fix issue #70. When certain characters were the first character in a line on a pocket label, the application would generate an exception. Appears to be an issue with the .indexOf function. If the .indexOf function returns a value less than 0, try the InStr function instead and adjust the offset into the string appropriately.